+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changes
 * Removed official support for Python 2.7, 3.4, 3.5 and 3.6, and added official
   support for Python 3.9, 3.10 and 3.11.
 
-* Removed ``SplashJsonResponse.body_as_unicode()``, replaced by
+* Deprecated ``SplashJsonResponse.body_as_unicode()``, to be replaced by
   ``SplashJsonResponse.text``.
 
 * Removed calls to obsolete ``to_native_str``, removed in Scrapy 2.8.